+++ internal/codec/writer.go
@@ -14,7 +14,7 @@ import (
 
 // WriterDocs is a static field documentation for output codecs.
 var WriterDocs = docs.FieldCommon(
-	"codec", "The way in which the bytes of messages should be written out into the output file. It's possible to write lines using a custom delimiter with the `delim:x` codec, where x is the character sequence custom delimiter.", "lines", "delim:\t", "delim:foobar",
+	"codec", "The way in which the bytes of messages should be written out into the output data stream. It's possible to write lines using a custom delimiter with the `delim:x` codec, where x is the character sequence custom delimiter.", "lines", "delim:\t", "delim:foobar",
 ).HasAnnotatedOptions(
 	"all-bytes", "Write the message to the file in full. If the file already exists the old content is deleted.",
 	"append", "Append messages to the file.",